rename and move authProvider+authStatusPolling to authn/ dir
